// src/app/modules/transportation/models/vehicle-details-view.ts
import { AttachmentViewModel } from 'app/modules/academics/student-admissions/model/attachment-view-model';

/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets VehicleDetailsView
 */
export interface VehicleDetailsView { 
    /**
     * Gets or Sets Id
     */
    id?: string;
    /**
     * Gets Or Sets
     */
    code?: string;
    /**
     * Gets Or Sets
     */
    name?: string;
    /**
     * Gets Or Sets
     */
    model?: string;
    /**
     * Gets Or Sets
     */
    manufacturer?: string;
    /**
     * Gets Or Sets
     */
    manufactureYear?: Date;
    /**
     * Gets Or Sets
     */
    regDate?: Date;
    /**
     * Gets Or Sets
     */
    regExpDate?: Date;
    /**
     * Gets Or Sets
     */
    regNumber?: string;
    /**
     * Gets Or Sets
     */
    engineNumber?: string;
    /**
     * Gets Or Sets
     */
    chassisNumber?: string;
    /**
     * Gets Or Sets
     */
    purchaseDate?: Date;
    /**
     * Gets Or Sets
     */
    purchaseCost?: number;
    /**
     * Gets Or Sets
     */
    purchaseOdometerReading?: number;
    /**
     * Gets Or Sets
     */
    maxCapacity?: number;
    /**
     * Gets Or Sets
     */
    photoAttachment?: AttachmentViewModel;
    /**
     * Gets Or Sets
     */
    registrationCertAttachment?: AttachmentViewModel;
}